Move source breakpoint/string conversion to util
